requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0                # via flask
cycler==0.10.0            # via matplotlib
flask-cors==3.0.8         # via -r requirements.in
flask==1.1.1              # via -r requirements.in, flask-cors
gunicorn==20.0.4          # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
joblib==0.14.1            # via scikit-learn
kiwisolver==1.1.0         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.2         # via wordcloud
nltk==3.4.5               # via -r requirements.in
numpy==1.18.1             # via matplotlib, pandas, scikit-learn, scipy, wordcloud
pandas==0.25.3            # via -r requirements.in
pillow==7.0.0             # via wordcloud
pyparsing==2.4.6          # via matplotlib
python-dateutil==2.8.1    # via matplotlib, pandas
pytz==2019.3              # via pandas
scikit-learn==0.22.1      # via -r requirements.in
scipy==1.4.1              # via scikit-learn
six==1.14.0               # via flask-cors, nltk, python-dateutil
werkzeug==0.16.0          # via flask
wordcloud==1.6.0          # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
